+++ Disasmo/Constants.cs
@@ -3,6 +3,4 @@
 public static class Constants
 {
     public const string DefaultJit = "clrjit.dll";
-    public const string Crossgen = "crossgen2.dll (R2R)";
-    public const string Ilc = "ilc (NativeAOT)";
 }