+++ src/Vsix/Properties/Settings.Designer.cs
@@ -61,8 +61,8 @@ namespace Disasmo.Properties {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("DOTNET_JitDiffableDasm=0;;DOTNET_ReadyToRun=1;;DOTNET_TieredPGO_InstrumentOnlyHot" +
-            "Code=0;;DOTNET_TC_CallCountingDelayMs=0")]
+        [global::System.Configuration.DefaultSettingValueAttribute("DOTNET_ReadyToRun=1;;DOTNET_TieredPGO_InstrumentOnlyHotCode=0;;DOTNET_TC_CallCoun" +
+            "tingDelayMs=0")]
         public string CustomEnvVars3_V15 {
             get {
                 return ((string)(this["CustomEnvVars3_V15"]));
@@ -207,13 +207,13 @@ namespace Disasmo.Properties {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute(@"-o:nativeaotbin.obj;;-r:%DOTNET_REPO%\artifacts\bin\coreclr\windows.x64.Checked\aotsdk\*.dll;;-r:%DOTNET_REPO%\artifacts\bin\microsoft.netcore.app.runtime.win-x64\Release\runtimes\win-x64\lib\net8.0\*.dll;;-g;;-O;;--initassembly:System.Private.CoreLib;;--initassembly:System.Private.StackTraceMetadata;;--initassembly:System.Private.TypeLoader;;--initassembly:System.Private.Reflection.Execution;;--directpinvokelist:%DOTNET_REPO%\artifacts\bin\coreclr\windows.x64.Checked\build\WindowsAPIs.txt;;--directpinvoke:System.Globalization.Native;;--directpinvoke:System.IO.Compression.Native;;--stacktracedata;;--scanreflection;;--feature:System.Runtime.Serialization.EnableUnsafeBinaryFormatterSerialization=false;;--feature:System.Resources.ResourceManager.AllowCustomResourceTypes=false;;--feature:System.Linq.Expressions.CanCompileToIL=false;;--feature:System.Linq.Expressions.CanEmitObjectArrayDelegate=false;;--feature:System.Linq.Expressions.CanCreateArbitraryDelegates=false;;--feature:System.Diagnostics.Tracing.EventSource.IsSupported=false;;")]
-        public string IlcArgs_V8 {
+        [global::System.Configuration.DefaultSettingValueAttribute(@"-o:nativeaotbin.obj;;-r:%DOTNET_REPO%\artifacts\bin\coreclr\windows.x64.Checked\aotsdk\*.dll;;-r:%DOTNET_REPO%\artifacts\bin\microsoft.netcore.app.runtime.win-x64\Release\runtimes\win-x64\lib\net9.0\*.dll;;-g;;-O;;--initassembly:System.Private.CoreLib;;--initassembly:System.Private.StackTraceMetadata;;--initassembly:System.Private.TypeLoader;;--initassembly:System.Private.Reflection.Execution;;--directpinvokelist:%DOTNET_REPO%\artifacts\bin\coreclr\windows.x64.Checked\build\WindowsAPIs.txt;;--directpinvoke:System.Globalization.Native;;--directpinvoke:System.IO.Compression.Native;;--stacktracedata;;--scanreflection;;--feature:System.Runtime.Serialization.EnableUnsafeBinaryFormatterSerialization=false;;--feature:System.Resources.ResourceManager.AllowCustomResourceTypes=false;;--feature:System.Linq.Expressions.CanCompileToIL=false;;--feature:System.Linq.Expressions.CanEmitObjectArrayDelegate=false;;--feature:System.Linq.Expressions.CanCreateArbitraryDelegates=false;;--feature:System.Diagnostics.Tracing.EventSource.IsSupported=false;;")]
+        public string IlcArgs_V9 {
             get {
-                return ((string)(this["IlcArgs_V8"]));
+                return ((string)(this["IlcArgs_V9"]));
             }
             set {
-                this["IlcArgs_V8"] = value;
+                this["IlcArgs_V9"] = value;
             }
         }
         
@@ -262,6 +262,18 @@ namespace Disasmo.Properties {
             }
             set {
                 this["OverridenTFM"] = value;
+            }
+        }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("False")]
+        public bool Diffable {
+            get {
+                return ((bool)(this["Diffable"]));
+            }
+            set {
+                this["Diffable"] = value;
             }
         }
     }